% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_cor.r
\name{CorrAR}
\alias{CorrAR}
\title{Covariance structure: Autoregressive}
\usage{
CorrAR(p, rho)
}
\arguments{
\item{p}{Dimension.}

\item{rho}{Correlation strength.}
}
\value{
A covariance matrix (p by p).
}
\description{
This function generates a covariance matrix with autoregressive structure.
}
\examples{
library(NRRR)
CorrAR(10, 0.5)
}
