% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_cor.r
\name{CorrCS}
\alias{CorrCS}
\title{Covariance structure: Compound symmetry}
\usage{
CorrCS(p, rho)
}
\arguments{
\item{p}{Dimension.}

\item{rho}{Correlation strength.}
}
\value{
A covariance matrix (p by p).
}
\description{
This function generates a covariance matrix with compound symmetry structure.
}
\examples{
library(NRRR)
CorrCS(10, 0.5)
}
