% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestRRR.cv.select_rcpp.R
\name{NestRRR.cv.select_rcpp}
\alias{NestRRR.cv.select_rcpp}
\title{NRRR with cross validation to select ranks (Rcpp)}
\usage{
NestRRR.cv.select_rcpp(
  Y,
  X,
  nfold = 10,
  norder = NULL,
  jx,
  jy,
  p,
  d,
  n,
  maxiter = 300,
  conv = 1e-04,
  method = c("RRR", "RRS")[1],
  lambda = 0,
  ic = c("BIC", "BICP", "AIC", "GCV")[1],
  dimred = c(TRUE, TRUE, TRUE),
  rankfix = NULL,
  xrankfix = NULL,
  yrankfix = NULL
)
}
\arguments{
\item{Y}{Response matrix.}

\item{X}{Design matrix.}

\item{nfold}{Number of folds, default is 10.}

\item{norder}{Assign samples to multiple folds for cross validation.}

\item{jx}{Number of basis functions to expand x(s).}

\item{jy}{Number of basis functions to expand y(t).}

\item{p}{Number of predictors.}

\item{d}{Number of responses.}

\item{n}{Sample size.}

\item{maxiter}{Maximum iteration number, Default 300.}

\item{conv}{Tolerance level to control convergence, default 1e-4.}

\item{method}{RRR (default) no ridge penalty; RRS use ridge penalty.}

\item{lambda}{Tuning parameter for the ridge penalty, only used when
method='RRS', default=0.}

\item{ic}{Information criterion, selected from BIC, BICP, AIC, GCV.}

\item{dimred}{A vector to decide whether do dimension reduction on certain
dimensions, with c(TRUE,TRUE,TRUE) corresponding to c(r,rx,ry),
TRUE (default) do dimension reduction; FALSE do not.
If dimred[1]=FALSE, r is provided by rankfix or min(jx*p,jy*d,rank(X));
If dimred[2]=FALSE, rx is provided by xrankfix or p;
If dimred[3]=FALSE, ry is provided by yrankfix or d.}

\item{rankfix}{Provide a value for r when dimred[1]=FALSE, default is NULL
which leads to min(jx*p,jy*d,rank(X)).}

\item{xrankfix}{Provide a value for rx when dimred[2]=FALSE, default is NULL
which leads to p.}

\item{yrankfix}{Provide a value for ry when dimred[3]=FALSE, default is NULL
which leads to d.}
}
\value{
The returned items are
  \item{Ag}{estimated matrix U.}
  \item{Bg}{estimated matrix V.}
  \item{Al}{estimated matrix A.}
  \item{Bl}{estimated matrix B.}
  \item{C}{estimated matrix C.}
  \item{df}{a scalar, estimated degrees of freedom of the selected model.}
  \item{sse}{a scalar, the sum of squared errors of the selected model.}
  \item{ic}{a vector contains values of BIC,BICP,AIC,GCV of the selected model.}
  \item{obj}{a vector contains all objective function (sse) values along
          iterations of the selected model.}
  \item{rx_path}{a matrix displays the path of selecting rx with cross validation.}
  \item{ry_path}{a matrix displays the path of selecting ry with cross validation.}
  \item{r_path}{a matrix displays the path of selecting r with cross validation.}
  \item{rank}{a scalar, the estimated r.}
  \item{rx}{a scalar, the estimated rx.}
  \item{ry}{a scalar, the estimated ry.}
  \item{rxErrmat}{a matrix displays errors in the path of selecting rx.
                  For each element, 0 indicates no error.}
  \item{ryErrmat}{a matrix displays errors in the path of selecting ry.
                  For each element, 0 indicates no error.}
  \item{rErrmat}{a matrix displays errors in the path of selecting r.
                  For each element, 0 indicates no error.}
}
\description{
This function applies Cross-Validation
 to select the optimal (r, rx, ry) with
 the blockwise coordinate descent algorithm.
}
\examples{
library(NRRR)
set.seed(1)
# Simulation setting 1 in NRRR paper
simDat <- nrrr.sim(n=100,ns=60,nt=60,r=5,rx=3,ry=3,
                   jx=8,jy=8,p=10,d=10,s2n=1,rho_X=0.5,
                   rho_E=0,Sigma=CorrAR)
fit_init <- with(simDat, NestRRR.cv.select_rcpp(Yest,Xest,nfold=10,norder=NULL,
                              jx=8,jy=8,p=10,d=10,n=100,
                              maxiter=300,conv=1e-4,
                              method=c('RRR','RRS')[1],lambda=0,
                              ic=c("BIC","BICP","AIC","GCV")[1],
                              dimred = c(TRUE,TRUE,TRUE),
                              rankfix=NULL,xrankfix=NULL,yrankfix=NULL))
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
