% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_NRRR.r
\name{NestRRR.prediction}
\alias{NestRRR.prediction}
\title{Generate prediction of response curves}
\usage{
NestRRR.prediction(tseq, X, sseq, Ag, Bg, Al, Bl, phi)
}
\arguments{
\item{tseq}{A sequence of the prediction time points of y(t).}

\item{X}{X(s), an array of observed predictor curves.}

\item{sseq}{A sequence of the observed time points of x(s).}

\item{Ag}{Global low-dimensional structure U, a d by ry matrix of rank ry.}

\item{Bg}{Global low-dimensional structure V, a p by rx matrix of rank rx.}

\item{Al}{Local low-dimensional structure A, a jy*ry by r matrix of rank r.}

\item{Bl}{Local low-dimensional structure B, a jx*rx by r matrix of rank r.}

\item{phi}{The set of basis functions to expand x(s).}
}
\value{
The returned items
  \item{Cstar}{\eqn{(U \otimes I_jy)A* B*^T(V \otimes I_jx)^T}, the coefficient matrix
          in equation (7).}
  \item{Jpsihalf}{psi is the set of basis functions to expand y(t); Jpsi is
          the covariance matrix of psi, and Jpsihalf is (Jpsi)^(1/2).}
  \item{Ypred}{the predicted response curves Ypred(t), an array of
          dimension (n, d, nt).}
}
\description{
This function generates prediction of response curves.
}
\examples{
library(NRRR)
set.seed(3)
# Simulation setting 2 in NRRR paper
simDat <- nrrr.sim(n=100,ns=100,nt=100,r=3,rx=3,ry=3,
                   jx=8,jy=8,p=20,d=20,s2n=1,rho_X=0.5,
                   rho_E=0,Sigma=CorrAR)
fit_nrrr <- with(simDat, NestRRR.select(Yest,Xest,Ag0=NULL,Bg0=NULL,
                              jx=8,jy=8,p=20,d=20,n=100,
                              maxiter=300,conv=1e-4,quietly=FALSE,
                              method=c('RRR','RRS')[1],lambda=0,
                              ic=c("BIC","BICP","AIC","GCV")[1],
                              dimred = c(TRUE,TRUE,TRUE),rankfix=NULL))
Ypred_nrrr <- NestRRR.prediction(simDat$tseq,simDat$X,simDat$sseq,
                                 fit_nrrr$Ag,fit_nrrr$Bg,fit_nrrr$Al,fit_nrrr$Bl,
                                 simDat$phi)
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
