% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestRRR.select.r
\name{NestRRR.select}
\alias{NestRRR.select}
\title{Select ranks with information criterion}
\usage{
NestRRR.select(
  Y,
  X,
  Ag0 = NULL,
  Bg0 = NULL,
  jx,
  jy,
  p,
  d,
  n,
  maxiter = 300,
  conv = 1e-04,
  quietly = FALSE,
  method = c("RRR", "RRS")[1],
  lambda = 0,
  ic = c("BIC", "BICP", "AIC", "GCV")[1],
  dimred = c(TRUE, TRUE, TRUE),
  rankfix = NULL
)
}
\arguments{
\item{Y}{Response matrix with n rows and jy*d columns.}

\item{X}{Design matrix with n rows and jx*p columns.}

\item{Ag0}{Initial estimator of U, if NULL then generate it by
function \code{NestRRRini()}.}

\item{Bg0}{Initial estimator of V, if NULL then generate it by
function \code{NestRRRini()}.}

\item{jx}{Number of basis functions to expand x(s).}

\item{jy}{Number of basis functions to expand y(t).}

\item{p}{Number of predictors.}

\item{d}{Number of responses.}

\item{n}{Sample size.}

\item{maxiter}{Maximum iteration number of the blockwise coordinate descent algorithm, default 300.}

\item{conv}{Tolerance level to control convergence, default 1e-4.}

\item{quietly}{FALSE (default) show the rank selection process, if TRUE then not.}

\item{method}{RRR (default) no ridge penalty; RRS use ridge penalty.}

\item{lambda}{Tuning parameter for the ridge penalty, only used when method='RRS', default=0.}

\item{ic}{Information criterion, selected from BIC, BICP, AIC, GCV.}

\item{dimred}{Vector to decide whether do dimension reduction on certain
dimensions, with c(TRUE,TRUE,TRUE) corresponding to c(r,rx,ry),
TRUE (default) do dimension reduction; FALSE do not.
If dimred[1]=FALSE, r is provided by rankfix or min(jx*p,jy*d, rank(X));
If dimred[2]=FALSE, rx equals to p;
If dimred[3]=FALSE, ry equals to d.}

\item{rankfix}{Provide a value for r when dimred[1]=FALSE, default is NULL
which leads to min(jx*p,jy*d,rank(X))}
}
\value{
The returned results containing
  \item{Ag}{the global low-dimensional structure U, a d by ry matrix of rank ry.}
  \item{Bg}{the global low-dimensional structure V, a p by rx matrix of rank rx.}
  \item{Al}{the local low-dimensional structure A.}
  \item{Bl}{the local low-dimensional structure B.}
  \item{C}{the NRRR estimator of the coefficient matrix C.}
  \item{df}{a scalar, the estimated degrees of freedom of the NRRR model.}
  \item{sse}{a scalar, the sum of squared errors of the selected model.}
  \item{ic}{a vector contains values of BIC,BICP,AIC,GCV of the selected model.}
  \item{obj}{ a vector contains all objective function (sse) values along iterations of the selected model.}
  \item{rank}{a scalar, the estimated r.}
  \item{rx}{a scalar, the estimated rx.}
  \item{ry}{a scalar, the estimated ry.}
}
\description{
This function select the optimal (r, rx, ry) with a specified information criterion
based on the proposed blockwise coordinate descent algorithm.
}
\examples{
library(NRRR)
set.seed(3)
# Simulation setting 2 in NRRR paper
simDat <- nrrr.sim(n=100,ns=100,nt=100,r=3,rx=3,ry=3,
                   jx=8,jy=8,p=20,d=20,s2n=1,rho_X=0.5,
                   rho_E=0,Sigma=CorrAR)
fit_init <- with(simDat, NestRRR.select(Yest,Xest,Ag0=NULL,Bg0=NULL,
                              jx=8,jy=8,p=20,d=20,n=100,
                              maxiter=300,conv=1e-4,quietly=FALSE,
                              method=c('RRR','RRS')[1],lambda=0,
                              ic=c("BIC","BICP","AIC","GCV")[1],
                              dimred = c(TRUE,TRUE,TRUE),rankfix=NULL))
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
