% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestRRRini.r
\name{NestRRRini}
\alias{NestRRRini}
\title{Generate initial estimators for NRRR}
\usage{
NestRRRini(Y, X, r, rx, ry, jx, jy, p, d, n)
}
\arguments{
\item{Y}{Response matrix with n rows and jy*d columns.}

\item{X}{Design matrix with n rows and jx*p columns.}

\item{r}{Rank of the local reduced-rank structure.}

\item{rx}{Number of latent predictors.}

\item{ry}{Number of latent responses.}

\item{jx}{Number of basis functions to expand x(s).}

\item{jy}{Number of basis functions to expand y(t).}

\item{p}{Number of predictors.}

\item{d}{Number of responses.}

\item{n}{Sample size.}
}
\value{
The returned items
  \item{Ag}{the global low-dimensional structure U, a d by ry matrix of rank ry.}
  \item{Bg}{the global low-dimensional structure V, a p by rx matrix of rank rx.}
}
\description{
This function estimates U and V to initialize the blockwise coordinate descent algorithm.
}
\examples{
library(NRRR)
simDat <- nrrr.sim(n=100,ns=200,nt=200,r=5,rx=3,ry=3,
                   jx=15,jy=15,p=10,d=6,s2n=1,rho_X=0.5,
                   rho_E=0,Sigma=CorrAR)
fit_init <- with(simDat, NestRRRini(Y=Yest,X=Xest,r=5,
                 rx=3,ry=3,jx=15,jy=15,p=10,d=6,n=100))
fit_init$Ag
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
