% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_NRRR.r
\name{NRRR.sim}
\alias{NRRR.sim}
\title{Generate simulation data}
\usage{
NRRR.sim(n, ns, nt, r, rx, ry, jx, jy, p, d,
         s2n, rho_X, rho_E, Sigma = "CorrAR")
}
\arguments{
\item{n}{sample size.}

\item{ns}{number of time points at which the predictor trajectory is observed.}

\item{nt}{number of time points at which the response trajectory is observed.}

\item{r}{rank of the local reduced-rank structure.}

\item{rx}{number of latent predictors.}

\item{ry}{number of latent responses.}

\item{jx}{number of basis functions to expand the predictor trajectory.}

\item{jy}{number of basis functions to expand the response trajectory.}

\item{p}{number of predictors.}

\item{d}{number of responses.}

\item{s2n}{a positive number to specify the signal to noise ratio.}

\item{rho_X}{a scalar between 0 and 1 to specify the correlation strength among covariates.}

\item{rho_E}{a scalar between 0 and 1 to specify the correlation strength among random errors.}

\item{Sigma}{the correlation structure. Two options are available,
"CorrAR": autoregressive,
"CorrCS": compound symmetry.}
}
\value{
The function returns a list:
  \item{Ag}{the matrix U.}
  \item{Bg}{the matrix V.}
  \item{Al}{the matrix A.}
  \item{Bl}{the matrix B.}
  \item{C}{the coefficient matrix C.}
  \item{Alstar}{A* in equation (7) of the NRRR paper.}
  \item{Blstar}{B* in equation (7) of the NRRR paper.}
  \item{Cstar}{\eqn{(U \otimes I_jy)A* B*^T(V \otimes I_jx)^T} in equation (7) of the NRRR paper.}
  \item{tseq}{a sequence of time points at which the response trajectory is observed.}
  \item{psi}{the set of basis functions to expand the response trajectory.}
  \item{Jpsi}{the correlation matrix of psi. Jpsihalf is (Jpsi)^(1/2).}
  \item{sseq}{a sequence of time points at which the predictor trajectory is observed.}
  \item{phi}{the set of basis functions to expand the predictor trajectory.}
  \item{Jphi}{the correlation matrix of phi.}
  \item{E}{the random error matrix.}
  \item{Y}{an array of dimension \code{(n, d, nt)}, i.e., the generated response observations.}
  \item{X}{an array of dimension \code{(n, p, ns)}, i.e., the generated predictor observations.}
  \item{Ytrue}{an array of dimension \code{(n, d, nt)}. The response observations without random errors.}
  \item{Yest}{the response matrix used in NRRR estimation, and is of dimension n-by-jy*d.}
  \item{Xest}{the design matrix used in NRRR estimation, and is of dimension n-by-jx*p.}
}
\description{
This function generates simulation data for the nested reduced-rank
regression under the multivariate functional regression scenario.
With user-specified model dimensions and ranks,
this function first generates functional multivariate predictors and
responses and then transform the functional regression problem into a
conventional finite-dimensional regression problem through basis expansion
and truncation. The B-spline basis is used to conduct basis expansion.
}
\examples{
library(NRRR)
simDat <- NRRR.sim(n = 100, ns = 200, nt = 200, r = 5, rx = 3, ry = 3,
                   jx = 15, jy = 15, p = 10, d = 6, s2n = 1, rho_X = 0.5,
                   rho_E = 0, Sigma = "CorrAR")
simDat$Ag
}
