% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.r
\name{Obj}
\alias{Obj}
\title{Compute || Y - XC ||_F^2}
\usage{
Obj(Y, X, Ag, Bg, Al, Bl, jx, jy)
}
\arguments{
\item{Y}{Response matrix with n rows and jy*d columns.}

\item{X}{Design matrix with n rows and jx*p columns.}

\item{Ag}{Matrix U.}

\item{Bg}{Matrix V.}

\item{Al}{Matrix A.}

\item{Bl}{Matrix B.}

\item{jx}{Number of basis functions to expand x(s).}

\item{jy}{Number of basis functions to expand y(t).}
}
\value{
The returned items
  \item{C}{the NRRR estimator of the coefficient matrix C.}
  \item{XC}{a matrix, prediction of Y.}
  \item{E}{a matrix, estimated error matrix.}
  \item{sse,obj}{ a scaler, the sum of squared errors.}
}
\description{
This function computes the sum of squared errors
with given (A,B,U,V) and (X,Y).
}
