% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRR.r
\name{RRR}
\alias{RRR}
\title{RRR with fixed rank}
\usage{
RRR(
  Y,
  X,
  nrank = 1,
  weight = FALSE,
  Gamma = diag(ncol(Y)),
  ypy.svd = TRUE,
  c.svd = FALSE
)
}
\arguments{
\item{Y}{Response matrix.}

\item{X}{Design matrix.}

\item{nrank}{The given rank.}

\item{weight}{If TRUE, then weighted criterion is performed.}

\item{Gamma}{A weight matrix.}

\item{ypy.svd}{If TRUE, svd function is used.}

\item{c.svd}{If TRUE, output the svd of the coefficient matrix.}
}
\value{
The returned results containing
  \item{C}{the estimated matrix C.}
  \item{C_ls}{the least square estimator.}
  \item{rank}{the rank value.}
}
\description{
This function computes reduced rank regression estimator with a given rank.
}
