% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRRRcv_NRRR.r
\name{RRRR.cv}
\alias{RRRR.cv}
\title{RRS with cross validation to select rank and tuning parameter}
\usage{
RRRR.cv(
  Y,
  X,
  nfold = 10,
  rankmax = min(dim(Y), dim(X)),
  nlam = 100,
  lambda = seq(0, 100, length = nlam),
  norder = NULL,
  nest.tune = FALSE,
  fold.drop = 0
)
}
\arguments{
\item{Y}{Response matrix.}

\item{X}{Design matrix.}

\item{nfold}{Number of folds.}

\item{rankmax}{The maximum rank.}

\item{nlam}{Number of tuning parameters.}

\item{lambda}{Tuning sequence.}

\item{norder}{Assign samples into different folds.}

\item{nest.tune}{Whether tune the rank and lambda in a nested way.}

\item{fold.drop}{Number of folds to drop.}
}
\value{
The returned results containing
  \item{cr_path}{a matrix display the path of model selection.}
  \item{C}{the estimated low-rank coefficient matrix.}
  \item{rank}{the selected rank value.}
  \item{lam}{the selected tuning parameter for ridge penalty.}
}
\description{
This function performs reduced rank ridge regression with model selected by
 cross validation.
}
\references{
Mukherjee, A., & Zhu, J. (2011).
Reduced Rank Ridge Regression and Its Kernel Extensions.
Statistical analysis and data mining,
4(6), 612–622.
}
