% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRRR-package.R
\docType{package}
\name{NRRR-package}
\alias{NRRR}
\alias{NRRR-package}
\title{NRRR: Nested Reduced-Rank Regression}
\description{
The nested reduced-rank regression (NRRR) approach can be used in
     fitting a regression model with multivariate functional responses and 
     predictors to achieve tailored dimension reduction. A two-level 
     low-rank structure is imposed on the functional regression surfaces: a 
     global low-rank structure identifies latent responses and predictors and a 
     local low-rank structure controls the complexity of the association between 
     the latent variables. The applicability of NRRR goes beyond the functional 
     setup, for example, it can also be applied to vector autoregressive 
     modeling of time series and tensor-on-tensor regression.
}
\details{
A package to implement Nested reduced-rank regression.
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
\keyword{internal}
