% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRRR-package.R
\docType{package}
\name{NRRR-package}
\alias{NRRR}
\alias{NRRR-package}
\title{NRRR: Multivariate Functional Regression via Nested Reduced-Rank Regularization}
\description{
The multivariate functional regression with a nested reduced-rank structure proposed by Liu et al. (2020) <arXiv:2003.04786>.
}
\details{
This package provides the functions to fit a multivariate functional regression where a nested reduced-rank structure is imposed on the regression surface to control the model complexity.
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv:2003.04786 [stat.ME].
}
\keyword{internal}
