% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRRR-package.R
\docType{package}
\name{NRRR-package}
\alias{NRRR}
\alias{NRRR-package}
\title{NRRR: Nested Reduced-Rank Regression}
\description{
An implementation of the nested reduced-rank regression (NRRR). 
    It includes the functions to do estimation, model selection and prediction.
}
\details{
A package to implement Nested reduced-rank regression.
}
\references{
Liu, X., Ma, S., & Chen, K. (2020).
Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
\keyword{internal}
