% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRRR.plot.Pred.r
\name{NRRR.plot.pred}
\alias{NRRR.plot.pred}
\title{Plot the predicted response trajectory}
\usage{
NRRR.plot.pred(Ypred, Y = NULL, i_ind = 1, yi_ind = 1,
               tseq, t_index = NULL, x_lab = NULL,
               y_lab = NULL)
}
\arguments{
\item{Ypred}{an array of dimension \code{(n, d, length(tseq))} where n is
the sample size and d is the number of components in the
multivariate response. The response trajectory is predicted at
a set of time points \code{tseq}.}

\item{Y}{an array of dimension \code{(n, d, length(tseq))}. It is the truly
observed response trajectory in the form of discrete observations at
time points \code{tseq}. This term is only available
when the function is applied to a testing set.}

\item{i_ind}{a user-specified sample ID for which the plot is to be drawn.
If the prediction is only conducted on one sample,
then set \code{i_ind = 1}; otherwise it should be an integer
less than or equal to n. Default is 1.}

\item{yi_ind}{a user-specified response component index for which the plot
is to be drawn. It should be an integer between 0 and d.
Default is 1.}

\item{tseq}{a sequence of time points at which the predicted response values
are obtained.}

\item{t_index}{the user-specified x-axis tick marks, and it should be
given as a vector of character strings of the same length as tseq.
Default is NULL.}

\item{x_lab, y_lab}{the user-specified x-axis and y-axis label,
and it should be given as a character string, e.g., x_lab = "Time".
Default is NULL.}
}
\value{
A ggplot2 object. If the observed response trajectory is available, then
        it is plotted as a black line while the predicted trajectory is in red.
}
\description{
This function plots the predicted response trajectory based on the output of \code{\link{NestRRR.prediction}}.
}
\details{
An examples of its usage can be found in the vignette of electricity demand analysis.
}
\references{
Liu, X., Ma, S., & Chen, K. (2020). Multivariate Functional Regression via Nested Reduced-Rank Regularization.
arXiv: Methodology.
}
\author{
Xiaokang Liu and Kun Chen
}
