% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_NRRR.r
\name{nrrr.sim}
\alias{nrrr.sim}
\title{Generate simulation data}
\usage{
nrrr.sim(n, ns, nt, r, rx, ry, jx, jy, p, d, s2n, rho_X, rho_E, Sigma = CorrAR)
}
\arguments{
\item{n}{Sample size.}

\item{ns}{Number of discrete observations of x(s).}

\item{nt}{Number of discrete observations of y(t).}

\item{r}{Rank.}

\item{rx}{Number of latent predictors.}

\item{ry}{Number of latent responses.}

\item{jx}{Number of basis functions to expand x(s).}

\item{jy}{Number of basis functions to expand y(t).}

\item{p}{Number of predictors.}

\item{d}{Number of responses.}

\item{s2n}{Signal-to-noise ratio.}

\item{rho_X}{The correlation strength among columns of X,
a scalar between 0 and 1.}

\item{rho_E}{The correlation strength among columns of E,
a scalar between 0 and 1.}

\item{Sigma}{The correlation structure, select from
CorrAR and CorrCS.}
}
\value{
The generated data containing
  \item{Ag}{the global low-dimensional structure U, a d by ry matrix of rank ry.}
  \item{Bg}{the global low-dimensional structure V, a p by rx matrix of rank rx.}
  \item{Al}{the local low-dimensional structure A, a (jy*ry) by r matrix of rank r.}
  \item{Bl}{the local low-dimensional structure B, a (jx*rx) by r matrix of rank r.}
  \item{C}{the coefficient matrix C.}
  \item{Alstar}{A* in equation (7).}
  \item{Blstar}{B* in equation (7).}
  \item{Cstar}{\eqn{(U \otimes I_jy)A* B*^T(V \otimes I_jx)^T} in equation (7), the coefficient matrix
           used in data generation.}
  \item{tseq}{a sequence of the observed time points of y(t) of length nt.}
  \item{psi}{the set of basis functions to expand y(t).}
  \item{Jpsi}{the correlation matrix of psi, and Jpsihalf is (Jpsi)^(1/2).}
  \item{sseq}{a sequence of the observed time points of x(s) of length ns.}
  \item{phi}{the set of basis functions to expand x(s).}
  \item{Jphi}{the correlation matrix of phi.}
  \item{E}{the random error matrix.}
  \item{Y}{Y(t), an array of dimension (n, d, nt), with random error.}
  \item{X}{X(s), an array of dimension (n, p, ns).}
  \item{Ytrue}{Ytrue(t), an array of fimension (n, d, nt), without random error.}
  \item{Yest}{response matrix of dimension n by (jy*d), used in estimation.}
  \item{Xest}{design matrix of dimension n by (jx*p), used in estimation.}
}
\description{
This function generates simulation data for the nested reduced-rank
functional regression model.
}
\examples{
library(NRRR)
simDat <- nrrr.sim(n=100,ns=200,nt=200,r=5,rx=3,ry=3,
                   jx=15,jy=15,p=10,d=6,s2n=1,rho_X=0.5,
                   rho_E=0,Sigma=CorrAR)
simDat$Ag
}
